refac: ReservationValidator 제거 및 검증로직 Reservation 에서 처리하도록 리팩토링
